Reorganize utils module and fix import paths
